+++ resources/views/back/managers/index.blade.php
@@ -16,26 +16,36 @@
 
 @section('content')
 
-    @if($managers->isEmpty() && $modelManager->can('create'))
-        <div class="center-center stack-xl">
-            <a href="{{ $modelManager->route('create') }}" class="btn btn-primary squished-l">
-                <i class="icon icon-zap icon-fw"></i> Tijd om een {{ $modelManager->details()->singular }} toe te voegen
-            </a>
-        </div>
-    @endif
-
-    @if((($managers->isEmpty() && !collect(request()->query())->isEmpty()) || (!$managers->isEmpty())) && (!$modelManager::filters()->isEmpty() || $modelManager::sections()->has('sidebar')))
-        <div class="row gutter stack">
+    @if($managers->isEmpty() && !$modelManager->filters()->anyApplied())
+        @if($modelManager->can('create'))
+            <div class="center-center stack-xl">
+                <a href="{{ $modelManager->route('create') }}" class="btn btn-primary squished-l">
+                    <i class="icon icon-zap icon-fw"></i> Tijd om een {{ $modelManager->details()->singular }} toe te voegen
+                </a>
+            </div>
+        @else
+            <p>Nog geen {{ $modelManager->details()->singular }} toegevoegd. Je hebt echter niet de nodige rechten om er een toe te voegen.</p>
+        @endif
+    @else
+        <div class="row gutter">
             <div class="column-9">
                 <div class="row gutter-s">
-                    @foreach($managers as $manager)
-                        @include('chief::back.managers._partials._rowitem')
-                        @include('chief::back.managers._partials.delete-modal')
+                    @if($managers->isEmpty())
+                        <div class="stack">
+                            <p>Geen resultaten gevonden voor uw huidige filtering.
+                            <br><a href="{{ $modelManager->route('index') }}">Bekijk alle resultaten</a>
+                            </p>
+                        </div>
+                    @else
+                        @foreach($managers as $manager)
+                            @include('chief::back.managers._partials._rowitem')
+                            @include('chief::back.managers._partials.delete-modal')
 
-                        @if($manager->isAssistedBy('archive'))
-                            @include('chief::back.managers._partials.archive-modal')
-                        @endif
-                    @endforeach
+                            @if($manager->isAssistedBy('archive'))
+                                @include('chief::back.managers._partials.archive-modal')
+                            @endif
+                        @endforeach
+                    @endif
                 </div>
             </div>
             <div class="column-3">
@@ -62,26 +72,6 @@
 
             </div>
         </div>
-    @else
-        <div class="row gutter stack">
-            @foreach($managers as $manager)
-                @include('chief::back.managers._partials._rowitem')
-                @include('chief::back.managers._partials.delete-modal')
-                @if($manager->isAssistedBy('archive'))
-                    @include('chief::back.managers._partials.archive-modal')
-                @endif
-            @endforeach
-        </div>
-
-        {!! $modelManager->sections()->sidebar !!}
-
-        @if($modelManager->isAssistedBy('archive') && $archiveAssistant = $modelManager->assistant('archive'))
-            @if( ! $archiveAssistant->findAll()->isEmpty())
-                <div class="center-x">
-                    <a class="squished-s" href="{{ $archiveAssistant->route('index') }}">Bekijk de gearchiveerde items</a>
-                </div>
-            @endif
-        @endif
     @endif